move add_global_connector
