+++ qtgui/sliders/__init__.py
@@ -5,3 +5,4 @@ __all__ = []
 from .base import *
 from .local_sliders import *
 from .global_sliders import *
+from .add_global_connector import AddGlobalConnectorWindow
