+++ qtgui/sliders/add_global_connector.py
